+++ federation/tests/conftest.py
@@ -4,6 +4,7 @@ import pytest
 
 from federation.entities.diaspora.entities import DiasporaPost
 from federation.tests.fixtures.keys import get_dummy_private_key
+from federation.tests.fixtures.payloads import DIASPORA_PUBLIC_PAYLOAD
 
 
 @pytest.fixture(autouse=True)
@@ -23,6 +24,11 @@ def disable_network_calls(monkeypatch):
 
 
 @pytest.fixture
+def diaspora_public_payload():
+    return DIASPORA_PUBLIC_PAYLOAD
+
+
+@pytest.fixture
 def diasporapost():
     return DiasporaPost()
 
@@ -30,3 +36,8 @@ def diasporapost():
 @pytest.fixture
 def private_key():
     return get_dummy_private_key()
+
+
+@pytest.fixture
+def public_key(private_key):
+    return private_key.publickey().exportKey()